rename the io.py module to config.py
