+++ src/dfastbe/bank_erosion/data_models.py
@@ -9,7 +9,7 @@ from shapely.geometry import LineString, Point
 from geopandas.geoseries import GeoSeries
 from dfastio.xyc.models import XYCModel
 from dfastbe.io.data_models import BaseRiverData, BaseSimulationData
-from dfastbe.io.io import ConfigFile, log_text
+from dfastbe.io.config import ConfigFile, log_text
 
 
 GenericType = TypeVar("GenericType")
