+++ src/dfastbe/bank_lines/data_models.py
@@ -3,7 +3,7 @@ from typing import List, Tuple
 from shapely.geometry import LineString, MultiLineString, Point
 from shapely.geometry.polygon import Polygon
 
-from dfastbe.io.io import log_text
+from dfastbe.io.config import log_text
 from dfastbe.io.data_models import BaseRiverData, BaseSimulationData, LineGeometry
 
 MAX_RIVER_WIDTH = 1000
@@ -170,7 +170,7 @@ class BankLinesRiverData(BaseRiverData):
 
         Examples:
             ```python
-            >>> from dfastbe.io.io import ConfigFile
+            >>> from dfastbe.io.config import ConfigFile
             >>> config_file = ConfigFile.read("tests/data/bank_lines/meuse_manual.cfg")
             >>> bank_lines_river_data = BankLinesRiverData(config_file)
             No message found for read_chainage
@@ -214,7 +214,7 @@ class BankLinesRiverData(BaseRiverData):
 
         Examples:
             ```python
-            >>> from dfastbe.io.io import ConfigFile
+            >>> from dfastbe.io.config import ConfigFile
             >>> from unittest.mock import patch
             >>> config_file = ConfigFile.read("tests/data/bank_lines/meuse_manual.cfg")
             >>> bank_lines_river_data = BankLinesRiverData(config_file)  # doctest: +ELLIPSIS
