+++ src/dfastbe/io/data_models.py
@@ -5,7 +5,7 @@ import netCDF4
 from shapely.geometry import LineString, Point
 from shapely import prepare
 from geopandas.geodataframe import GeoDataFrame
-from dfastbe.io.io import SimulationFilesError, ConfigFile, get_bbox, log_text
+from dfastbe.io.config import SimulationFilesError, ConfigFile, get_bbox, log_text
 
 
 __all__ = ["BaseSimulationData", "BaseRiverData", "LineGeometry"]
