+++ tests/test_bank_lines/test_bank_lines.py
@@ -10,7 +10,7 @@ from shapely.geometry import LineString, MultiLineString, Polygon
 from dfastbe.bank_lines.bank_lines import BankLines
 from dfastbe.cmd import run
 from dfastbe.io.data_models import BaseSimulationData
-from dfastbe.io.io import ConfigFile
+from dfastbe.io.config import ConfigFile
 
 matplotlib.use('Agg')
