+++ tests/test_erosion/test_bank_erosion.py
@@ -5,7 +5,7 @@ from unittest.mock import MagicMock
 
 from dfastbe.bank_erosion.bank_erosion import Erosion
 from dfastbe.cmd import run
-from dfastbe.io.io import ConfigFile
+from dfastbe.io.config import ConfigFile
 from dfastbe.bank_erosion.data_models import FairwayData
 
 matplotlib.use('Agg')
